backend/requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
efficientnet_pytorch
scikit-learn
numpy
